Swapping two variables using c
Call by value:
We can observe that even when we change the content of x and y in the scope of the swap function, these changes does not reflect on x and y variables defined in the scope of main. This is because we call swap() by value and it will get separate memory for x and y so the changes made in swap() will not reflect in main().
Call by reference:
We can observe in function parameters instead of using int x,int y we used int *x,int *y and in function call instead of giving x,y we give &x,&y this methodology is call by reference as we used pointers as function parameter which will get original parameters address instead of their value. & operator is used to give address of the variables and * is used to access the memory location that pointer is pointing. As the function variable is pointing to same memory location as original parameter the changes made in swap() reflect in main() which we could see in the above output.